refactor!: rename the module to wits and settle the monorepo layout
The module is github.com/TheDonDope/wits. A terminal interface was the
first thing built, not the whole of what this is: a server and a web
interface are next, and naming the module after one of three components
would have aged badly the moment the second arrived.

One module with several commands, rather than one per component. The
domain packages are shared, and a module boundary between a server and the
ledger it serves would mean tagging and versioning the ledger against
itself on every change. Angular brings its own toolchain, so wits-ui sits
beside cmd/ rather than inside it.

  wits/
    cmd/wits/           the terminal interface and the commands
    cmd/wits-server/    the REST API                        (planned)
    pkg/                the domain, shared by both
    wits-ui/            the web interface, in Angular       (planned)

The binary stays `wits`. It is what gets typed daily and what every
example documents, and cmd/wits-tui producing a `wits-tui` binary would
have bought a tidier directory listing at the cost of the command's name.

cmd/wits/home held one command in a package of its own and now sits with
the rest as commands.Home. The Makefile builds ./cmd/wits rather than
./cmd/wits/main.go, so a second file in that package cannot be silently
left out of a release build.

Behaviour is unchanged: the workbook imports to the same 1369 entries and
`wits status` is identical, byte for byte, to before the rename.

BREAKING CHANGE: the import path is now github.com/TheDonDope/wits.
