+++ .chglog/config.yml
@@ -2,7 +2,7 @@ style: github
 template: CHANGELOG.tpl.md
 info:
   title: CHANGELOG
-  repository_url: https://github.com/TheDonDope/wits-tui
+  repository_url: https://github.com/TheDonDope/wits
 options:
   commits:
     # filters:
