+++ cmd/wits/commands/grind.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"fmt"
 
-	"github.com/TheDonDope/wits-tui/pkg/journal"
+	"github.com/TheDonDope/wits/pkg/journal"
 	"github.com/spf13/cobra"
 )
