+++ cmd/wits/commands/log.go
@@ -5,7 +5,7 @@ import (
 	"text/tabwriter"
 	"time"
 
-	"github.com/TheDonDope/wits-tui/pkg/journal"
+	"github.com/TheDonDope/wits/pkg/journal"
 	"github.com/spf13/cobra"
 )
