--- cmd/wits/home/doc.go
@@ -1,2 +1,0 @@
-// Package home provides the main entry point for the wits application
-package home // import "github.com/TheDonDope/wits-tui/cmd/wits/home"
