+++ pkg/cannabis/doc.go
@@ -1,2 +1,2 @@
 // Package cannabis provides all data shapes for the different information regarding the cannabis plant.
-package cannabis // import "github.com/TheDonDope/wits-tui/pkg/cannabis"
+package cannabis // import "github.com/TheDonDope/wits/pkg/cannabis"
