+++ pkg/catalog/catalog.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	can "github.com/TheDonDope/wits-tui/pkg/cannabis"
+	can "github.com/TheDonDope/wits/pkg/cannabis"
 	"gopkg.in/yaml.v3"
 )
