+++ pkg/tui/format.go
@@ -7,7 +7,7 @@ import (
 
 	"charm.land/lipgloss/v2"
 
-	"github.com/TheDonDope/wits-tui/pkg/journal"
+	"github.com/TheDonDope/wits/pkg/journal"
 )
 
 // glyphs give each event type a shape, so the kind of an entry is legible
