+++ pkg/version/doc.go
@@ -1,2 +1,2 @@
 // Package version provides build information about the application
-package version // import "github.com/TheDonDope/wits-tui/pkg/version"
+package version // import "github.com/TheDonDope/wits/pkg/version"
